update: 보상 정보 갱신
- 식사는 레시피에 해당하는 아이템의 rank * 7 만큼 골드 획득 및 경험치 1 획득

+++ Assets/Scripts/Data/QuestData.cs
@@ -22,7 +22,7 @@ public class QuestData
     public int id;
     public int rank;
     public string title;
-    public float reward;
+    public int reward;
     public string rawNeedSpecs; // "str, dex, int, hp"
     public int time; // 소요 일수
     public string rawDropItems; // "itemId:probability,itemId:probability,..."
